Refactoring Submission file into separate classes to allow for upcoming Delivery classes
