+++ prime-router/src/main/kotlin/DetailedSubmissionHistory.kt
@@ -1,17 +1,12 @@
 package gov.cdc.prime.router
 
-import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
-import com.fasterxml.jackson.annotation.JsonInclude
-import com.fasterxml.jackson.annotation.JsonInclude.Include
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.annotation.JsonPropertyOrder
 import com.fasterxml.jackson.annotation.JsonValue
 import com.microsoft.azure.functions.HttpStatus
-import gov.cdc.prime.router.azure.WorkflowEngine
 import gov.cdc.prime.router.azure.db.enums.TaskAction
 import java.time.OffsetDateTime
-import java.util.UUID
 
 /**
  * A `DetailedSubmissionHistory` represents the detailed life history of a submission of a message from a sender.
@@ -32,51 +27,29 @@ import java.util.UUID
 )
 class DetailedSubmissionHistory(
     @JsonProperty("submissionId")
-    val actionId: Long,
-    @JsonIgnore
-    val actionName: TaskAction,
-    @JsonProperty("timestamp")
-    val createdAt: OffsetDateTime,
-    val httpStatus: Int? = null,
-    @JsonIgnore
-    var reports: MutableList<DetailReport>?,
-    @JsonIgnore
-    var logs: List<DetailActionLog> = emptyList()
+    actionId: Long,
+    actionName: TaskAction,
+    createdAt: OffsetDateTime,
+    httpStatus: Int? = null,
+    reports: MutableList<DetailReport>?,
+    logs: List<DetailActionLog> = emptyList()
+) : DetailedReportFileHistory(
+    actionId,
+    actionName,
+    createdAt,
+    httpStatus,
+    reports,
+    logs,
 ) {
-    /**
-     * The report ID.
-     */
-    var id: String? = null
-
     /**
      * The destinations.
      */
     var destinations = mutableListOf<Destination>()
 
     /**
-     * Error log for the submission.
-     */
-    val errors = mutableListOf<ConsolidatedActionLog>()
-
-    /**
-     * Warning log for the submission.
-     */
-    val warnings = mutableListOf<ConsolidatedActionLog>()
-
-    /**
-     * The schema topic.
-     */
-    var topic: String? = null
-
-    /**
      * The sender of the input report.
      */
     var sender: String? = null
-
-    /**
-     * The input report's external name.
-     */
-    var externalName: String? = null
 
     /**
      * The step in the delivery process for a submission
@@ -125,26 +98,11 @@ class DetailedSubmissionHistory(
     var actualCompletionAt: OffsetDateTime? = null
 
     /**
-     * The number of warnings.  Note this is not the number of consolidated warnings.
-     */
-    val warningCount = logs.count { it.type == ActionLogLevel.warning }
-
-    /**
-     * The number of errors.  Note this is not the number of consolidated errors.
-     */
-    val errorCount = logs.count { it.type == ActionLogLevel.error }
-
-    /**
      * Number of destinations that actually had/will have data sent to.
      */
     val destinationCount: Int get() {
         return destinations.filter { it.itemCount != 0 }.size
     }
-
-    /**
-     * Number of report items.
-     */
-    var reportItemCount: Int? = null
 
     init {
         reports?.forEach { report ->
@@ -182,8 +140,6 @@ class DetailedSubmissionHistory(
                 topic = report.schemaTopic
             }
         }
-        errors.addAll(consolidateLogs(ActionLogLevel.error))
-        warnings.addAll(consolidateLogs(ActionLogLevel.warning))
     }
 
     fun enrichWithDescendants(descendants: List<DetailedSubmissionHistory>) {
@@ -299,47 +255,6 @@ class DetailedSubmissionHistory(
     }
 
     /**
-     * Consolidate the [logs] filtered by an optional [filterBy] action level, so to list similar messages once
-     * with a list of items they relate to.
-     * @return the consolidated list of logs
-     */
-    internal fun consolidateLogs(filterBy: ActionLogLevel? = null):
-        List<ConsolidatedActionLog> {
-        val consolidatedList = mutableListOf<ConsolidatedActionLog>()
-
-        // First filter the logs and sort by the message.  This first sorting can take care of sorting old messages
-        // that contain index numbers like "Report 3: xxxx"
-        val filteredList = when (filterBy) {
-            null -> logs
-            else -> logs.filter { it.type == filterBy }
-        }.sortedBy { it.detail.message }
-        // Now order the list so that logs contain first non-item messages, then item messages, and item messages
-        // are sorted by index.
-        val orderedList = (
-            filteredList.filter { it.scope != ActionLogScope.item }.sortedBy { it.scope } +
-                filteredList.filter { it.scope == ActionLogScope.item }.sortedBy { it.index }
-            ).toMutableList()
-        // Now consolidate the list
-        while (orderedList.isNotEmpty()) {
-            // Grab the first log.
-            val consolidatedLog = ConsolidatedActionLog(orderedList.first())
-            orderedList.removeFirst()
-            // Now find similar logs and consolidate it.  Note by using the iterator we can delete on the fly.
-            with(orderedList.iterator()) {
-                forEach { log ->
-                    if (consolidatedLog.canBeConsolidatedWith(log)) {
-                        consolidatedLog.add(log)
-                        remove()
-                    }
-                }
-            }
-            consolidatedList.add(consolidatedLog)
-        }
-
-        return consolidatedList
-    }
-
-    /**
      * Runs the calculations for the overallStatus field so that it can be done during init.
      * @returns The status from the Status enum that matches the current Submission state.
      */
@@ -424,190 +339,4 @@ class DetailedSubmissionHistory(
 
         return sentReports.plus(downloadedReports).maxWithOrNull(compareBy { it.createdAt })?.createdAt
     }
-}
-
-/**
- * Consolidated action log class to be output to the API JSON response.
- * @param log the base log message to be consolidated
- */
-@JsonInclude(Include.NON_NULL)
-class ConsolidatedActionLog(log: DetailActionLog) {
-    /**
-     * The scope of the log.
-     */
-    val scope: ActionLogScope
-
-    /**
-     * The list of indices for item logs. An index can be null if there was no index provided with the log.
-     */
-    val indices: MutableList<Int?>?
-
-    /**
-     * The list of tracking IDs for item logs. A tracking ID can be null if there was no ID provided with the log.
-     */
-    val trackingIds: MutableList<String?>?
-
-    /**
-     * The log level.
-     */
-    @JsonIgnore
-    val type: ActionLogLevel
-
-    /**
-     * The field mapping for item logs.
-     */
-    val field: String?
-
-    /**
-     * The log message.
-     */
-    val message: String
-
-    init {
-        scope = log.scope
-        type = log.type
-        message = log.detail.message
-        if (log.detail.scope == ActionLogScope.item) {
-            field = if (log.detail is ItemActionLogDetail) log.detail.fieldMapping else null
-            indices = mutableListOf()
-            trackingIds = mutableListOf()
-        } else {
-            indices = null
-            trackingIds = null
-            field = null
-        }
-        add(log)
-    }
-
-    /**
-     * Add an action detail [log] to this consolidated log.
-     */
-    fun add(log: DetailActionLog) {
-        check(message == log.detail.message)
-        if (indices != null && trackingIds != null) {
-            indices.add(log.index)
-            trackingIds.add(log.trackingId)
-        }
-    }
-
-    /**
-     * Tests if a detail action log [other] can be consolidated into this existing consolidated log.
-     * @return true if the log can be consolidated, false otherwise
-     */
-    fun canBeConsolidatedWith(other: DetailActionLog): Boolean {
-        return this.message == other.detail.message && this.scope == other.scope && this.type == other.type
-    }
-}
-
-/**
- * Detail action log class used to read the data from the database.
- */
-@JsonInclude(Include.NON_NULL)
-@JsonIgnoreProperties(ignoreUnknown = true)
-class DetailActionLog(
-    val scope: ActionLogScope,
-    @JsonIgnore
-    val reportId: UUID?,
-    val index: Int?,
-    val trackingId: String?,
-    val type: ActionLogLevel,
-    val detail: ActionLogDetail,
-)
-
-@JsonInclude(Include.NON_NULL)
-@JsonIgnoreProperties(ignoreUnknown = true)
-data class DetailReport(
-    val reportId: UUID,
-    @JsonIgnore
-    val receivingOrg: String?,
-    @JsonIgnore
-    val receivingOrgSvc: String?,
-    @JsonIgnore
-    val sendingOrg: String?,
-    @JsonIgnore
-    val sendingOrgClient: String?,
-    @JsonIgnore
-    val schemaTopic: String?,
-    val externalName: String?,
-    val createdAt: OffsetDateTime?,
-    val nextActionAt: OffsetDateTime?,
-    val itemCount: Int,
-    @JsonIgnore
-    val itemCountBeforeQualFilter: Int?,
-)
-
-/**
- * Response use for the API for the filtered report items. This removes unneeded properties that exist in
- * ReportStreamFilterResult. ReportStreamFilterResult is used to serialize and deserialize to/from the database.
- * @param filterResult the filter result to use
- */
-data class ReportStreamFilterResultForResponse(@JsonIgnore private val filterResult: ReportStreamFilterResult) {
-    val filterType = filterResult.filterType
-    val filterName = filterResult.filterName
-    val filteredTrackingElement = filterResult.filteredTrackingElement
-    val filterArgs = filterResult.filterArgs
-    val message = filterResult.message
-}
-
-@JsonPropertyOrder(
-    value = [
-        "organization", "organizationId", "service", "itemCount", "itemCountBeforeQualFilter", "sendingAt"
-    ]
-)
-@JsonIgnoreProperties(ignoreUnknown = true)
-data class Destination(
-    @JsonProperty("organization_id")
-    val organizationId: String,
-    val service: String,
-    val filteredReportRows: List<String>?,
-    val filteredReportItems: List<ReportStreamFilterResultForResponse>?,
-    @JsonProperty("sending_at")
-    @JsonInclude(Include.NON_NULL)
-    val sendingAt: OffsetDateTime?,
-    val itemCount: Int,
-    @JsonProperty("itemCountBeforeQualityFiltering")
-    val itemCountBeforeQualFilter: Int?,
-    var sentReports: MutableList<DetailReport> = mutableListOf(),
-    var downloadedReports: MutableList<DetailReport> = mutableListOf(),
-) {
-    val organization: String?
-        get() = WorkflowEngine.settingsProviderSingleton.findOrganizationAndReceiver(
-            "$organizationId.$service"
-        )?.let { (org, _) ->
-            org.description
-        }
-}
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-class SubmissionHistory(
-    @JsonProperty("submissionId")
-    val actionId: Long,
-    @JsonProperty("timestamp")
-    val createdAt: OffsetDateTime,
-    @JsonProperty("sender")
-    val sendingOrg: String,
-    val httpStatus: Int,
-    @JsonInclude(Include.NON_NULL)
-    val externalName: String? = "",
-    @JsonIgnore
-    val reportId: String? = null,
-    @JsonIgnore
-    val schemaTopic: String? = null,
-    @JsonIgnore
-    val itemCount: Int? = null
-) {
-    /**
-     * The report ID.
-     */
-    val id = reportId
-
-    /**
-     * The topic.
-     */
-    val topic = schemaTopic
-
-    /**
-     * The number of items in the report.
-     */
-    val reportItemCount = itemCount
 }
